Update emails branch with api-based site (#32)
* Clean up TODO

* Restructure code

* Update db helper to read from auth table

* Add password encryption

* Catch and display login errors properly

* Mimic transactional db write

* Remove test values + clean up
