+++ KinConnections/templates/index.html
@@ -49,7 +49,7 @@
 					<li><a class="nav-link scrollto" href="/home">Home</a></li>
 					<li><a class="nav-link scrollto" href="/home#about">How it Works</a></li>
 					<li><a class="nav-link scrollto" href="/guide">Connecting Guide</a></li>
-					<li><a class="nav-link scrollto" href="/search">Connectors</a></li>
+					<li><a class="nav-link scrollto" href="/connectors">Connectors</a></li>
 					<!-- if logged in -->
 					{% if session.get('email', None) %}
 					<li class="dropdown"><a href="/profile/{{session.get('name', None)}}"><span>Hi {{session.get('name', None)}}!</span> <i
