+++ KinConnections/templates/signup.html
@@ -1,6 +1,4 @@
- <!-- TODO add 'I am a Connector/Connectee' -->
-
-{% extends "kc_template.html" %}
+ {% extends "kc_template.html" %}
 {% block page_title %} Login {% endblock %}
 
 {% block breadcrumbs %}
@@ -169,7 +167,6 @@
                               </div>
                             </div>
 
-                            <!-- TODO "have you attended GE camps? Y/N" -> displays question for location, year, position (fac/participant)? -->
                             <div class="form-group mt-3">
                               <label for="attended_ge">Have you attended Global Encounters or CONNECT?</label><br/>
                               <input type="radio" name="attended_ge" value="yes" id="yes" onclick="javascript:displayGE();" required=""> Yes &nbsp;&nbsp;
